added comments

+++ src/controllers/Notices/noticesController.js
@@ -3,9 +3,10 @@ const Notice = require('../../models/Notice')
 // Create a notice
 const createNotice = async (req, res) => {
     try {
-
+        // Create a notice
         const newNotice = await Notice.createNotice(req.body);
 
+        // Return new notice data
         res.status(201).json(newNotice);
 
     } catch (error) {
@@ -17,8 +18,10 @@ const createNotice = async (req, res) => {
 // Get all notices
 const getAllNotices = async (req, res) => {
     try {
-
+        // Search for all notices
         const notices = await Notice.find();
+
+        // Return notice list
         res.json(notices);
 
     } catch (error) {
@@ -30,8 +33,10 @@ const getAllNotices = async (req, res) => {
 // Get a notice by ID
 const getNoticeById = async (req, res) => {
     try {
-
+        // Search for notice
         const notice = await Notice.getNoticeById(req.params.id);
+
+        // Return notice
         res.json(notice);
 
     } catch (error) {
@@ -43,7 +48,20 @@ const getNoticeById = async (req, res) => {
 // Update a notice by ID
 const updateNotice = async (req, res) => {
     try {
+        // Search for notice
+        const foundNotice = await Notice.findById(req.params.id)
 
+        // If notice not found throw error
+        if (!foundNotice) {
+            throw new Error('Notice not found');
+        }
+
+        // Only admin or committee or notice owner can update a notice
+        if ((req.user.role.role !== 'admin' || req.user.role.role !== 'committee') && req.user._id.toString() !== foundNotice.author._id.toString()) {
+            throw new Error('Unauthorized');
+        }
+
+        // Make notice update
         const updatedNotice = await Notice.updateNotice(req.params.id, req.body);
 
         res.json(updatedNotice);
@@ -57,9 +75,10 @@ const updateNotice = async (req, res) => {
 // Add a comment to a todo
 const addCommentNotice = async (req, res) => {
     try {
-
+        // Add comment to notice
         const updatedNotice = await Notice.addComment(req.params.id, req.body);
 
+        // Return updated notice
         res.json(updatedNotice);
 
     } catch (error) {
@@ -71,9 +90,10 @@ const addCommentNotice = async (req, res) => {
 // Remove a comment from a todo
 const removeCommentNotice = async (req, res) => {
     try {
-
+        // Remove comment from notice
         const updatedNotice = await Notice.removeComment(req.params.id, req.params.commentId, req.user._id);
 
+        // Return updated notice
         res.json(updatedNotice);
 
     } catch (error) {
@@ -86,7 +106,23 @@ const removeCommentNotice = async (req, res) => {
 const deleteNotice = async (req, res) => {
     try {
 
+        // Search for notice
+        const foundNotice = await Notice.findById(req.params.id)
+
+        // If notice not found throw error
+        if (!foundNotice) {
+            throw new Error('Notice not found');
+        }
+
+        // Only admin or committee or notice owner can update a notice
+        if ((req.user.role.role !== 'admin' || req.user.role.role !== 'committee') && req.user._id.toString() !== foundNotice.author._id.toString()) {
+            throw new Error('Unauthorized');
+        }
+
+        // Delete notice
         const deletedNotice = await Notice.deleteNotice(req.params.id);
+
+        // Return deleted notice
         res.json(deletedNotice);
 
     } catch (error) {
